Be extra careful when deleting orphans.
Avoid having an exception thrown in case the orphan instance somehow
gets deleted between detection and deletion.

+++ object_database/service_manager/ServiceManager.py
@@ -300,7 +300,8 @@ class ServiceManager(object):
         self.stopServices(orphans)
         with self.db.transaction():
             for instance in orphans:
-                instance.delete()
+                if instance.exists():
+                    instance.delete()
 
     @revisionConflictRetry
     def collectDeadHosts(self):
